refactor(stores): extract shared parsers + pagination clamp helper
Centralise the three identical copies of safeParseTags / safeParseMetadata
and the two identical copies of clampPagination into
src/server/stores/_parsers.ts. ClawStashDB, VersionStore, and SearchStore
now all import the same helpers instead of carrying their own copies.

Behaviour is bit-for-bit identical:
- safeParseTags: typeof check, JSON.parse, Array.isArray + string filter,
  catch returns []
- safeParseMetadata: typeof check, JSON.parse, object/non-array guard,
  catch returns {}
- clampPagination: page/limit integer guards, default fallback, offset
  computed from clamped page

Verification:
- tsc --noEmit: clean
- vitest: 107/107 passing (same as baseline)
- next build: succeeds, all routes generated

Net effect: -83 lines duplication, +46 lines shared module = -37 LoC.
